quater tones 2

+++ src/lilypond/msr.h
@@ -405,8 +405,11 @@ class msrMusicXMLNoteData
     char                fMusicXMLStep;
     bool                fMusicXMLStepIsARest;
     bool                fMusicXMLStepIsUnpitched;
-    
-    int                 fMusicXMLAlteration;
+
+    // The alter element represents chromatic alteration
+    // in number of semitones (e.g., -1 for flat, 1 for sharp).
+    // Decimal values like 0.5 (quarter tone sharp) are used for microtones.
+    float               fMusicXMLAlteration;
     
     int                 fMusicXMLOctave;
